requirements.txt
absl-py==2.3.1
addict==2.4.0
attrs==25.3.0
basicsr==1.4.2
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
clip==0.2.0
contourpy==1.3.2
cycler==0.12.1
decorator==4.4.2
exceptiongroup==1.3.0
facexlib==0.3.0
ffmpeg-python==0.2.0
filelock==3.18.0
filterpy==1.4.5
fonttools==4.59.0
fsspec==2025.7.0
ftfy==6.3.1
future==1.0.0
gfpgan==1.3.8
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
instaloader==4.14.2
intervaltree==3.1.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.7.3
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.5
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
open_clip_torch==2.32.0
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
proglog==0.1.12
protobuf==6.31.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.8.1
pyobjc-core==11.1
pyobjc-framework-ApplicationServices==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-CoreText==11.1
pyobjc-framework-Quartz==11.1
pyparsing==3.2.3
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
realesrgan==0.3.0
regex==2024.11.6
requests==2.32.4
rubicon-objc==0.5.2
safetensors==0.5.3
scenedetect==0.6.6
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
selenium==4.34.2
sentence-transformers==5.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.14.0
tag==0.5
tb-nightly==2.21.0a20250808
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.17
tkinterdnd2==0.4.3
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.2
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yapf==0.43.0
